Inizio implementazione provvisoria HeatFollow

+++ SportlightSA/src/main/java/it/uniroma3/radeon/sa/functions/SumReduceFunction.java
@@ -2,11 +2,11 @@ package it.uniroma3.radeon.sa.functions;
 
 import org.apache.spark.api.java.function.Function2;
 
-public class SumReduceFunction implements Function2<Integer, Integer, Integer> {
+public class SumReduceFunction implements Function2<Long, Long, Long> {
 	
 	private static final long serialVersionUID = 1L;
 
-	public Integer call(Integer a, Integer b) {
+	public Long call(Long a, Long b) {
 		return a + b;
 	}
